Fix LP64 crashes: COORDINATE, pathfinding, facing, icon sets, overlaps
- COORDINATE: typedef changed from unsigned long (8 bytes) to uint32_t
  (4 bytes) — fixes all coordinate packing/unpacking
- Desired_Facing8/256: implement from FACING8.ASM reference; Facing8
  returns DirType scale (0,32,64..224), not raw 0..7
- Find_Path overlap arrays: unsigned long → uint32_t for 32-bit bit
  indexing
- Get_Radar_Icon: disabled (returns NULL) — SHP/KeyFrame format
  mismatch causes heap corruption via Build_Frame
- HelpClass::Set_Text: disabled — crashes on Text_Overlap_List LP64
- Map::Read_Binary: read 2 raw bytes per cell (not sizeof enum struct)
- Init_Theater: null-check palette Retrieve result
- MixFileClass::Offset: skip nodes with null Buffer
- All debug fprintf converted to DBG() macro
- DBG traces throughout AI chain for regression debugging
- queue.cpp: COORD format specifiers %lx → %x for uint32_t

+++ modern.src/defines.h
@@ -1590,7 +1590,7 @@ typedef enum RadioMessageType {
 **	with cell resolution. The COORD type is used for map coordinates that
 **	have a lepton resolution.
 */
-typedef unsigned long	COORDINATE;
+typedef uint32_t	COORDINATE;  /* must be 32-bit: packed as [Xcell:8][Xlepton:8][Ycell:8][Ylepton:8] */
 typedef signed short		CELL;
 
 typedef unsigned short	TARGET;
@@ -2384,7 +2384,7 @@ typedef struct {
 	int				Cost;					// Accumulated terrain cost.
 	int				Length;				// Command string length.
 	FacingType		*Command;			// Pointer to command string.
-	unsigned long	*Overlap;			// Pointer to overlap list
+	uint32_t			*Overlap;			// Pointer to overlap list (32-bit for bit indexing)
 	CELL				LastOverlap;		// stores position of last overlap
 	CELL				LastFixup;			// stores position of last overlap
 } PathType;
